composer/requirements.txt
apache-airflow-providers-google
google-cloud-storage
google-cloud-bigquery
pyyaml
requests
dbt-bigquery
